Move the widget into its own file.
