+++ Widgets/Widgets.swift
@@ -218,150 +218,3 @@ extension UIImage {
     }
 
 }
-
-// MARK: - VPN Status Widget
-
-enum VPNStatus {
-    case status(NEVPNStatus)
-    case error
-    case notConfigured
-}
-struct VPNStatusTimelineEntry: TimelineEntry {
-    let date: Date
-    let status: VPNStatus
-    let location: String
-
-    internal init(date: Date, status: VPNStatus = .notConfigured, location: String = "No Location") {
-        self.date = date
-        self.status = status
-        self.location = location
-    }
-}
-
-class VPNStatusTimelineProvider: TimelineProvider {
-
-    typealias Entry = VPNStatusTimelineEntry
-
-    func placeholder(in context: Context) -> VPNStatusTimelineEntry {
-        return VPNStatusTimelineEntry(date: Date())
-    }
-
-    func getSnapshot(in context: Context, completion: @escaping (VPNStatusTimelineEntry) -> Void) {
-        let entry = VPNStatusTimelineEntry(date: Date())
-        completion(entry)
-    }
-
-    func getTimeline(in context: Context, completion: @escaping (Timeline<VPNStatusTimelineEntry>) -> Void) {
-        NETunnelProviderManager.loadAllFromPreferences { managers, error in
-            let expiration = Date().addingTimeInterval(TimeInterval.minutes(1))
-
-            if let error {
-                let entry = VPNStatusTimelineEntry(date: expiration, status: .error)
-                let timeline = Timeline(entries: [entry], policy: .atEnd)
-                completion(timeline)
-                return
-            }
-
-            guard let manager = managers?.first else {
-                let entry = VPNStatusTimelineEntry(date: expiration, status: .notConfigured)
-                let timeline = Timeline(entries: [entry], policy: .atEnd)
-                completion(timeline)
-                return
-            }
-
-            let status = manager.connection.status
-            let enabled = (status == .connected || status == .connecting)
-
-            let entry = VPNStatusTimelineEntry(date: expiration, status: .status(status))
-            let timeline = Timeline(entries: [entry], policy: .atEnd)
-
-            completion(timeline)
-        }
-    }
-}
-
-extension NEVPNStatus {
-    var description: String {
-        switch self {
-        case .connected:
-            return "Connected"
-        case .connecting:
-            return "Connecting"
-        case .disconnected:
-            return "Disconnected"
-        case .disconnecting:
-            return "Disconnecting"
-        case .invalid:
-            return "Invalid"
-        case .reasserting:
-            return "Reasserting"
-        default:
-            return "Unknown Status"
-        }
-    }
-}
-
-struct VPNStatusView: View {
-    @Environment(\.widgetFamily) var family: WidgetFamily
-    var entry: VPNStatusTimelineProvider.Entry
-
-    @ViewBuilder
-    var body: some View {
-        if #available(iOSApplicationExtension 17.0, *) {
-            VStack {
-                switch entry.status {
-                case .status(let status):
-                    Text(status.description)
-                    // Text("Location: \(entry.location)").font(.footnote)
-
-                    Spacer()
-
-                    switch status {
-                    case .connected:
-                        Button(intent: DisableVPNIntent()) {
-                            Text("Disconnect")
-                                .font(.body)
-                                .padding()
-                                .background(Color.black)
-                                .foregroundStyle(.white)
-                                .clipShape(Capsule())
-                        }
-                    case .disconnected:
-                        Button(intent: EnableVPNIntent()) {
-                            Text("Connect")
-                                .font(.body)
-                                .padding()
-                                .background(Color.black)
-                                .foregroundStyle(.white)
-                                .clipShape(Capsule())
-                        }
-                    default:
-                        Text("Changing status...")
-                    }
-                case .error:
-                    Text("Error")
-                case .notConfigured:
-                    Text("VPN Not Configured")
-                }
-            }
-            .containerBackground(for: .widget) {
-                Color.orange
-            }
-        } else {
-            Text("iOS 17 required")
-        }
-    }
-}
-
-struct VPNStatusWidget: Widget {
-    let kind: String = "VPNStatusWidget"
-
-    var body: some WidgetConfiguration {
-        StaticConfiguration(kind: kind, provider: VPNStatusTimelineProvider()) { entry in
-            VPNStatusView(entry: entry)
-        }
-        .configurationDisplayName("VPN Status")
-        .description("View and manage the DuckDuckGo VPN status")
-        .supportedFamilies([.systemSmall])
-    }
-}
